ScreenControllerの修正とButtonController系の統合
