+++ Assets/MyAssets/Prefab/UI/Button/TutorialButton.prefab
@@ -147,7 +147,7 @@ GameObject:
   - component: {fileID: 378742639810163582}
   - component: {fileID: 5041802051000355855}
   - component: {fileID: 6465256400554506799}
-  - component: {fileID: 1993470981213539984}
+  - component: {fileID: 1797065526586466887}
   m_Layer: 5
   m_Name: TutorialButton
   m_TagString: UI
@@ -274,7 +274,7 @@ Animator:
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorControllerStateOnDisable: 0
---- !u!114 &1993470981213539984
+--- !u!114 &1797065526586466887
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -283,9 +283,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 6567449633154523510}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: cd55b89c25bf0c54f9b98067b1874dc6, type: 3}
+  m_Script: {fileID: 11500000, guid: 6a9fa6554a981644782f707731f67d99, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   screenController: {fileID: 0}
   button: {fileID: 5041802051000355855}
-  sceneAnimator: {fileID: 0}
+  screenState: 1
