--- Assets/MyAssets/Script/GiveUpButtonController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 56dac7d808b60fd43941586dd6dc23be
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
