--- Assets/MyAssets/Script/TutorialButtonController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cd55b89c25bf0c54f9b98067b1874dc6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
