+++ Assets/Scenes/GameScene.unity
@@ -227,6 +227,18 @@ Animator:
     type: 3}
   m_PrefabInstance: {fileID: 1003112903}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &733639879 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 5037443923982805173, guid: 89116517d9e399545953c2a16e3dde80,
+    type: 3}
+  m_PrefabInstance: {fileID: 1003112903}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a1cbdc7c0ab6c5d40938d10d3630d77d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &1003112903
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -234,6 +246,11 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 743971479241198798, guid: 89116517d9e399545953c2a16e3dde80,
+        type: 3}
+      propertyPath: screenState
+      value: 2
+      objectReference: {fileID: 0}
     - target: {fileID: 967765959934542951, guid: 89116517d9e399545953c2a16e3dde80,
         type: 3}
       propertyPath: m_AnchorMax.y
@@ -381,7 +398,17 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3236909572267825189, guid: 89116517d9e399545953c2a16e3dde80,
         type: 3}
+      propertyPath: fadeOutTime
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 3236909572267825189, guid: 89116517d9e399545953c2a16e3dde80,
+        type: 3}
       propertyPath: screenState
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3236909572267825189, guid: 89116517d9e399545953c2a16e3dde80,
+        type: 3}
+      propertyPath: transitionTime
       value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 3236909572267825189, guid: 89116517d9e399545953c2a16e3dde80,
@@ -454,6 +481,11 @@ PrefabInstance:
       propertyPath: m_AnchoredPosition.y
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 4760659368937188665, guid: 89116517d9e399545953c2a16e3dde80,
+        type: 3}
+      propertyPath: tileTypeSelector
+      value: 
+      objectReference: {fileID: 733639879}
     - target: {fileID: 4853411174751323246, guid: 89116517d9e399545953c2a16e3dde80,
         type: 3}
       propertyPath: m_AnchorMax.y
